add board render & board marking

+++ .eslintrc.js
@@ -14,6 +14,6 @@ module.exports = {
     },
     "rules": {
         "semi": ["warn", "always"],
-        "quotes": ["warn", "double"]
+        "quotes": ["warn", "single"]
     }
 };
